Fall back to using spring's annotations for setup of jobRepo
Introduce locally defined repo definitions in the annotation for the applicable batch job
Some are read only and use H2, others read / write using POSTGRES

Also clean up imports in build file.
