+++ build.gradle.kts
@@ -108,10 +108,6 @@ dependencies {
   implementation("org.springframework.boot:spring-boot-starter-oauth2-resource-server:3.2.2")
   implementation("org.springframework.boot:spring-boot-starter-oauth2-client:3.2.2")
   implementation("com.nimbusds:oauth2-oidc-sdk:11.9.1")
-  implementation("org.springframework.boot:spring-boot-starter-webflux:3.2.2")
-  implementation("org.springframework.boot:spring-boot-starter-oauth2-resource-server:3.2.2")
-  implementation("org.springframework.boot:spring-boot-starter-oauth2-client:3.2.2")
-  implementation("com.nimbusds:oauth2-oidc-sdk:11.9.1")
 
   // database
   implementation("org.springframework.boot:spring-boot-starter-data-jpa")
